extract cluster-type crate
